Create next_msg_author variable
To correctly punish an non-responsive player, the TimeoutContract must
know which player is expected to send a message at a given time. And
although during gameplay this information is accurately represented by
`game-state`'s `current_player`, that is not always the case.

Therefore it was need to create a new variable to expose this
information to the webapp.