Leave queue immediately during onboarding
Was previously allowing the player to return to the Lobby to resume his
waiting. This is actually a little dangerous to the player, because it
would allow him to open two instances of the application and have both
of them receive the GameReady event. This way, both instances would try
to submit turn data, and this would lead to an invalid submission, and
consequently a challenge and loss of tokens.