Add exception handling for external messages
As a safety measure, we’re adding try/catch blocks to the libTMCG
methods that deal with data received from the opponent, as there’s no
practical way to validate if a payload is valid or not.

Exception handling on the engine wasn’t a thing until now because it
adds a severe overhead for the WASM version, but now that the native
engine is the default this shouldn’t be a problem.